Broke learning agents out into its own package.
